+++ apps/earn/components/PoolSection/PoolComposition.tsx
@@ -1,11 +1,13 @@
-import { Native } from '@sushiswap/currency'
 import { formatUSD } from '@sushiswap/format'
 import { Pool } from '@sushiswap/client'
-import { AppearOnMount, Currency, Table, Typography } from '@sushiswap/ui'
-import { usePrices } from '@sushiswap/react-query'
-import { FC } from 'react'
+import { Currency } from '@sushiswap/ui'
+import React, { FC } from 'react'
+import { List } from '@sushiswap/ui/future/components/list/List'
 
-import { useGraphPool } from '../../lib/hooks'
+import { usePoolGraphData, useTokenAmountDollarValues } from '../../lib/hooks'
+import { ChainId } from '@sushiswap/chain'
+import { Native } from '@sushiswap/currency'
+import { usePrices } from '@sushiswap/react-query'
 
 interface PoolCompositionProps {
   pool: Pool
@@ -13,99 +15,54 @@ interface PoolCompositionProps {
 
 export const PoolComposition: FC<PoolCompositionProps> = ({ pool }) => {
   const { data: prices } = usePrices({ chainId: pool.chainId })
-  const {
-    data: { token0, token1, reserve0, reserve1, liquidityNative },
-  } = useGraphPool(pool)
+  const { data, isLoading } = usePoolGraphData({
+    poolAddress: pool.address,
+    chainId: pool.chainId as ChainId,
+  })
+  const fiatValues = useTokenAmountDollarValues({ chainId: pool.chainId, amounts: [data?.reserve0, data?.reserve1] })
 
   return (
-    <div className="flex flex-col w-full gap-4">
-      <div className="flex items-center justify-between px-2">
-        <Typography weight={600} className="dark:text-slate-50 text-gray-900">
-          Pool Composition
-        </Typography>
-        <AppearOnMount>
-          <Typography variant="sm" weight={400} className="text-gray-600 dark:text-slate-400 text-slate-600">
-            Total Assets:{' '}
-            <span className="font-semibold dark:text-slate-50 text-gray-900">
-              {' '}
-              {formatUSD(
-                (liquidityNative ?? 0) * Number(prices?.[Native.onChain(pool.chainId).wrapped.address]?.toFixed(10))
-              )}
-            </span>
-          </Typography>
-        </AppearOnMount>
+    <List>
+      <div className="flex items-center justify-between">
+        <List.Label>Pool Liquidity</List.Label>
+        <List.Label>
+          {formatUSD(
+            (data?.liquidityNative ?? 0) * Number(prices?.[Native.onChain(pool.chainId).wrapped.address]?.toFixed(10))
+          )}
+        </List.Label>
       </div>
-      <Table.container className="w-full">
-        <Table.table>
-          <Table.thead>
-            <Table.thr>
-              <Table.th>
-                <div className="text-left">Token</div>
-              </Table.th>
-              <Table.th>
-                <div className="text-left">Amount</div>
-              </Table.th>
-              <Table.th>
-                <div className="text-left">Value</div>
-              </Table.th>
-            </Table.thr>
-          </Table.thead>
-          <Table.tbody>
-            <Table.tr>
-              <Table.td>
-                <div className="flex items-center gap-3">
-                  <Currency.Icon currency={token0} width={24} height={24} />
-                  <Typography weight={600} variant="sm" className="dark:text-slate-50 text-gray-900">
-                    {token0.symbol}
-                  </Typography>
-                </div>
-              </Table.td>
-              <Table.td>
-                <Typography weight={500} variant="sm" className="text-gray-600 dark:text-slate-400 text-slate-600">
-                  {reserve0?.toSignificant(6)}
-                </Typography>
-              </Table.td>
-              <Table.td>
-                <AppearOnMount>
-                  <Typography weight={600} variant="sm" className="dark:text-slate-50 text-gray-900">
-                    {formatUSD(
-                      prices?.[token0.wrapped.address] && reserve0
-                        ? reserve0.multiply(prices?.[token0.wrapped.address].asFraction).toSignificant(6)
-                        : ''
-                    )}
-                  </Typography>
-                </AppearOnMount>
-              </Table.td>
-            </Table.tr>
-            <Table.tr>
-              <Table.td>
-                <div className="flex items-center gap-3">
-                  <Currency.Icon currency={token1} width={24} height={24} />
-                  <Typography weight={600} variant="sm" className="dark:text-slate-50 text-gray-900">
-                    {token1.symbol}
-                  </Typography>
-                </div>
-              </Table.td>
-              <Table.td>
-                <Typography weight={500} variant="sm" className="text-gray-600 dark:text-slate-400 text-slate-600">
-                  {reserve1?.toSignificant(6)}
-                </Typography>
-              </Table.td>
-              <Table.td>
-                <AppearOnMount>
-                  <Typography weight={600} variant="sm" className="dark:text-slate-50 text-gray-900">
-                    {formatUSD(
-                      prices?.[token1.wrapped.address] && reserve1
-                        ? reserve1.multiply(prices?.[token1.wrapped.address].asFraction).toSignificant(6)
-                        : ''
-                    )}
-                  </Typography>
-                </AppearOnMount>
-              </Table.td>
-            </Table.tr>
-          </Table.tbody>
-        </Table.table>
-      </Table.container>
-    </div>
+      <List.Control>
+        {data ? (
+          <List.KeyValue flex title={`${data.reserve0.currency.symbol}`}>
+            <div className="flex flex-col gap-2">
+              <div className="flex items-center gap-2">
+                <Currency.Icon currency={data.reserve0.currency} width={18} height={18} />
+                {data.reserve0.toSignificant(4)} {data.reserve0.currency.symbol}{' '}
+                <span className="text-gray-600 dark:text-slate-400">({formatUSD(fiatValues?.[0] || 0)})</span>
+              </div>
+            </div>
+          </List.KeyValue>
+        ) : isLoading ? (
+          <List.KeyValue skeleton />
+        ) : (
+          <></>
+        )}
+        {data ? (
+          <List.KeyValue flex title={`${data.reserve1.currency.symbol}`}>
+            <div className="flex flex-col gap-2">
+              <div className="flex items-center gap-2">
+                <Currency.Icon currency={data.reserve1.currency} width={18} height={18} />
+                {data.reserve1.toSignificant(4)} {data.reserve1.currency.symbol}{' '}
+                <span className="text-gray-600 dark:text-slate-400">({formatUSD(fiatValues?.[1] || 0)})</span>
+              </div>
+            </div>
+          </List.KeyValue>
+        ) : isLoading ? (
+          <List.KeyValue skeleton />
+        ) : (
+          <></>
+        )}
+      </List.Control>
+    </List>
   )
 }