+++ packages/tines/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @sushiswap/tines
 
+## 1.0.4
+
+### Patch Changes
+
+- Updated dependencies [813200e55]
+  - @sushiswap/currency@0.0.4
+
 ## 1.0.3
 
 ### Patch Changes