+++ packages/v2-sdk/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @sushiswap/v2-sdk
 
+## 0.0.3
+
+### Patch Changes
+
+- Updated dependencies [813200e55]
+  - @sushiswap/currency@0.0.4
+
 ## 0.0.2
 
 ### Patch Changes